+++ src/event-mocks/index.ts
@@ -1,2 +1,5 @@
 export * from './api-gateway';
 export * from './dynamodb-stream';
+export * from './sns';
+export * from './custom-authorizer';
+export * from './cloudformation';