+++ src/index.ts
@@ -1,7 +1,13 @@
 import { getDeployedUrl, getApiGatewayUrl, getOutput } from './stack-output-parser';
 import { StackOutputPlugin } from './stack-output';
 import { context } from './mock-signature';
-import { apiGatewayEvent, dynamoDBStreamEvent } from './event-mocks';
+import {
+  apiGatewayEvent,
+  dynamoDBStreamEvent,
+  snsEvent,
+  customAuthorizerEvent,
+  cloudFormationCustomResourceEvent
+} from './event-mocks';
 
 // Need to set it up this way so that there is a default export so that the serverless plugin works as well as the function exports (which can be imported into a project)
 
@@ -17,3 +23,6 @@ module.exports.getOutput = getOutput;
 module.exports.context = context;
 module.exports.apiGatewayEvent = apiGatewayEvent;
 module.exports.dynamoDBStreamEvent = dynamoDBStreamEvent;
+module.exports.snsEvent = snsEvent;
+module.exports.customAuthorizerEvent = customAuthorizerEvent;
+module.exports.cloudFormationCustomResourceEvent = cloudFormationCustomResourceEvent;