+++ typings/index.d.ts
@@ -1,0 +1,3 @@
+type NestedPartial<T> = {
+  [P in keyof T]?: NestedPartial<T[P]>;
+};