+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [2.0.2]&nbsp;&nbsp;(2020-01-24)
+
+### Added
+
+- Added mock event generator for API Gateway custom authorizer, CloudFormation custom resource, and SNS
+
+### Changed
+
+- Updated documentation to include event mocks
+
 ## [2.0.1]&nbsp;&nbsp;(2020-01-23)
 
 ### Added
@@ -98,6 +108,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), 
 - Clarify project intent and provide additional examples in README
 - Publish now happens from git tags instead of master merges
 
+[2.0.2]: https://github.com/manwaring/serverless-plugin-test-helper/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/manwaring/serverless-plugin-test-helper/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/manwaring/serverless-plugin-test-helper/compare/v1.3.1...v2.0.0
 [1.3.1]: https://github.com/manwaring/serverless-plugin-test-helper/compare/v1.3.0...v1.3.1